+++ webpack.config.js
@@ -12,7 +12,15 @@ module.exports = (env, options) => {
       use: ['raw-loader'],
     },
     {
-      test: /\.(sa|sc|c)ss$/,
+      test: /\.module.(sa|sc|c)ss$/,
+      use: [
+        'raw-loader',
+        'postcss-loader',
+        'sass-loader',
+      ],
+    },
+    {
+      test: /(?<!\.module)\.(sa|sc|c)ss$/,
       use: [
         {
           loader: MiniCssExtractPlugin.loader,
@@ -29,7 +37,7 @@ module.exports = (env, options) => {
         'postcss-loader',
         'sass-loader',
       ],
-    }
+    },
   ];
 
   // if (mode === 'production') {